+++ clients/common/misc/rocsolver_arguments.hpp
@@ -54,6 +54,7 @@ public:
     rocblas_int perf = 0;
     rocblas_int singular = 0;
     rocblas_int iters = 5;
+    rocblas_int alg_mode = 0;
     rocblas_int mem_query = 0;
     rocblas_int profile = 0;
     rocblas_int profile_kernels = 0;
@@ -112,6 +113,7 @@ public:
         to_consume.erase("batch_count");
         to_consume.erase("verify");
         to_consume.erase("iters");
+        to_consume.erase("alg_mode");
         to_consume.erase("mem_query");
         to_consume.erase("profile");
         to_consume.erase("profile_kernels");