+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Full documentation for rocSOLVER is available at the [rocSOLVER documentation](h
 
 * Application of a sequence of plane rotations to a given matrix
     - LASR
+* Algorithm selection APIs for hybrid computation
+* SVD of bidiagonal matrices routine:
+    - BDSQR now supports hybrid computation
+* SVD of general matrices routine:
+    - GESVD now supports hybrid computation
 
 ### Changed
 ### Removed